rename from migration-cli to cluster-to-redpanda-migration
