+++ src/go/cluster-to-redpanda-migration/cluster-to-redpanda-migration.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+set -x
+docker run -v $(pwd)/:/workspace/ localhost/cluster-to-redpanda-migration:latest $1
