+++ src/go/cluster-to-redpanda-migration/cmd/version/version.go
