+++ src/go/cluster-to-redpanda-migration/pkg/migration/cluster_spec.go
