+++ src/go/cluster-to-redpanda-migration/tests/data/00-console-rp_output.yaml
