+++ src/go/cluster-to-redpanda-migration/tests/data/01-console-admin-api-rp_output.yaml
