+++ src/go/cluster-to-redpanda-migration/tests/data/02-console-kafka-mtls-rp_output.yaml
