+++ src/go/cluster-to-redpanda-migration/tests/data/additional_configs_output.yaml
