+++ src/go/cluster-to-redpanda-migration/tests/data/complex_cluster_001.yaml
