+++ src/go/cluster-to-redpanda-migration/tests/data/simple_cluster_output.yaml
