+++ src/go/cluster-to-redpanda-migration/tests/data/simple_resources_listeners.yaml
