+++ src/go/cluster-to-redpanda-migration/tests/simple_cluster_test.go
@@ -10,7 +10,7 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"sigs.k8s.io/yaml"
 
-	"github.com/redpanda-data/redpanda/src/go/migration-cli/cmd"
+	"github.com/redpanda-data/redpanda/src/go/cluster-to-redpanda-migration/cmd"
 )
 
 var (
