--- src/go/migration-cli/migration-cli.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-set -x
-docker run -v $(pwd)/:/workspace/ localhost/migration-cli:latest $1
